flask_application/requirements.txt
alembic==1.12.1
attrs==23.2.0
certifi==2024.2.2
Charm-Crypto==0.50
charset-normalizer==3.3.2
click==8.1.7
dill==0.3.7
exceptiongroup==1.2.0
Flask==2.2.5
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.0.5
greenlet==3.0.3
hypothesis==6.99.11
idna==3.6
importlib-metadata==6.7.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.2.4
MarkupSafe==2.1.5
pairing==0.1.3
pyparsing==2.4.0
requests==2.31.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.28
typing_extensions==4.7.1
urllib3==2.0.7
Werkzeug==2.2.3
zipp==3.15.0
